l0-platform/requirements.txt
fastapi>=0.111.0
uvicorn>=0.29.0
httpx>=0.27.0
pyyaml>=6.0
